TEAMLOCK DOCKER
Creating a docker image with  Apache2 and mod_wsgi.
The image is based on the python:3.8 docker image.
Also trying to provide a docker-compose to try it "out of the box".

The current configuration listen on port 80.
This will evolve once everything is tested and working.

README.md update will come just after the docker creation is finished.